requirements.txt
websocket-client
pymongo